refactor: 의존성 역전으로 Order -> Table 패키지 단방향 구조로 변경
