+++ src/main/java/kitchenpos/table/application/TableValidator.java
@@ -1,0 +1,11 @@
+package kitchenpos.table.application;
+
+import kitchenpos.table.domain.OrderTable;
+import kitchenpos.table.domain.TableGroup;
+
+public interface TableValidator {
+
+    void validateChangeStatus(OrderTable orderTable);
+
+    void validateUngroup(TableGroup tableGroup);
+}
